refactor database operations into db class
